Fixed type incompatibility issue that could cause possible data loss

+++ src/main/java/com/afollestad/bridge/RequestBuilder.java
@@ -28,7 +28,7 @@ public final class RequestBuilder implements AsResultsExceptions, Serializable {
   int readTimeout;
   int currentRetryCount;
   int totalRetryCount;
-  int retrySpacingMs;
+  long retrySpacingMs;
   RetryCallback retryCallback;
   boolean cancellable = true;
   Object tag;
